+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## upcoming
 
+## 0.0.5
+- Support for providing options on construction: `new Stack( options )`
+- Added `options.breakOn` option which allows developers to set when a stack
+  should break. Note: Stack will ALWAYS break on Error.
+
 ## 0.0.4
 - Debug is now using pattern `module-name:file/path`